+++ crates/valence_protocol/src/packets/play/update_selected_slot_c2s.rs
@@ -3,5 +3,5 @@ use super::*;
 #[derive(Copy, Clone, Debug, Encode, Decode, Packet)]
 #[packet(id = packet_id::UPDATE_SELECTED_SLOT_C2S)]
 pub struct UpdateSelectedSlotC2s {
-    pub slot: u16,
+    pub slot: u8,
 }
